+++ frontend/src/_components/CopyToClipboard/index.js
@@ -1,0 +1,3 @@
+import { CopyToClipboardComponent } from './CopyToClipboard';
+
+export default CopyToClipboardComponent;
