+++ frontend/src/_ui/JSONTreeViewer/JSONNodeArray.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { JSONNode } from './JSONNode';
+
+const JSONTreeArrayNode = ({ data, path, ...restProps }) => {
+  const keys = [];
+
+  for (let i = 0; i < data.length; i++) {
+    keys.push(String(i));
+  }
+
+  return keys.map((key, index) => {
+    const currentPath = [...path, key];
+    const _currentNode = key;
+    const props = { ...restProps };
+    props.currentNode = _currentNode;
+
+    return <JSONNode key={`arr-${key}/${index}`} data={data[Number(key)]} path={currentPath} {...props} />;
+  });
+};
+
+export default JSONTreeArrayNode;
