+++ frontend/src/_ui/JSONTreeViewer/JSONNodeObject.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { JSONNode } from './JSONNode';
+
+const JSONTreeObjectNode = ({ data, path, ...restProps }) => {
+  const nodeKeys = Object.keys(data);
+
+  return nodeKeys.map((key, index) => {
+    const currentPath = [...path, key];
+    const _currentNode = key;
+    const props = { ...restProps };
+    props.currentNode = _currentNode;
+
+    return <JSONNode key={`obj-${key}/${index}`} data={data[key]} path={currentPath} {...props} />;
+  });
+};
+
+export default JSONTreeObjectNode;
